refactor:#21 페이지 폴더 분리
